+++ mobile/pubspec.yaml
@@ -47,6 +47,7 @@ dependencies:
   mixpanel_flutter: ^2.5.0
   flutter_secure_storage: ^9.2.4
   record: ^6.2.0
+  smooth_page_indicator: ^1.2.0+3
 
 dev_dependencies:
   flutter_test:
